時間を0.1秒ごとに減らすように変更

+++ BigTimer/Model/TimeModel.swift
@@ -9,7 +9,7 @@ import Foundation
 
 class TimerModel: ObservableObject{
     var displayedTimeFormat: TimeFormat = .min1
-    var setTime: Double = 0
-    var timeLeft: Double = 0
+    var setTime: Int = 0
+    var timeLeft: Int = 0
     var timerStatus: TimerStatus = .ready
 }
